compile example in GCC and remove warnings in clang

+++ Source/Libraries/Asi/RuntimeCommandEngine/CommandWithGenericParams.cpp
@@ -45,18 +45,18 @@ CommandWithGenericParams::EvaluateImpl(std::vector<std::string> const &tokens) c
 	if (true == std::get<0>(cwp_atis))
 	{
 		bool match = last_good_index_p1 != static_cast<int32_t>(tokens.size());
-		int32_t potential_good_index = last_good_index_p1 + 1;
-		for (; (potential_good_index < static_cast<int32_t>(tokens.size())) &&
+		uint32_t potential_good_index = static_cast<uint32_t>(last_good_index_p1) + 1U;
+		for (; (potential_good_index < tokens.size()) &&
 		       (potential_good_index - commandWithoutParams.GetSignatureLength() < genericParameters.size()) &&
 		       genericParameters[potential_good_index - commandWithoutParams.GetSignatureLength()]
 		           ->IsInExpectedDomain(tokens[potential_good_index]);
 		     ++potential_good_index)
 		{
 		}
-		match = potential_good_index == static_cast<int32_t>(tokens.size());
+		match = potential_good_index == tokens.size();
 		match =
 		    match && (tokens.size() == commandWithoutParams.GetSignatureLength() + genericParameters.size());
-		last_good_index_p1 = potential_good_index - 1;
+		last_good_index_p1 = static_cast<int32_t>(potential_good_index) - 1;
 		std::get<0>(output) = match;
 		std::get<1>(output) = last_good_index_p1;
 		if (match)
